requirements.txt
-i https://pypi.python.org/simple
argcomplete==1.9.3
awscli==1.16.140
awsebcli==3.15.0
boto3==1.9.130
botocore==1.12.130
cement==2.8.2
certifi==2019.3.9
cfn-flip==1.2.0
chardet==3.0.4
click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.1.2
django==2.2
djangorestframework==3.9.2
docutils==0.14
durationpy==0.5
future==0.16.0
gunicorn==19.9.0
hjson==3.0.1
idna==2.7
itypes==1.1.0
jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
markdown==3.1
markupsafe==1.1.1
openapi-codec==1.3.2
pathspec==0.5.9
pillow==6.0.0
placebo==0.9.0
psycopg2-binary==2.8.1
psycopg2==2.8.1
pyasn1==0.4.5
pypiwin32==223 ; sys_platform == 'win32' and python_version >= '3.6'
python-dateutil==2.6.1 ; python_version >= '2.7'
python-slugify==1.2.4
pytz==2019.1
pywin32==224
pyyaml==3.13
pyzbar==0.1.8
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1 ; python_version >= '3.4'
wcwidth==0.1.7
werkzeug==0.15.2
wheel==0.33.1
whitenoise==4.1.2
wsgi-request-logger==0.4.6
zappa==0.48.2
